Update Scanner.java
Currently, there isn't any support for `- var someVar = "someString" statements. They must be `someVar = "someString"`. This breaks compatibility with the official Jade library. So to remedy this issue, a RegEx replace statement has been added that looks for the `- var` statements, removes them, and lets the Scanner continue on normally.